[修正] UXを改善した
