+++ lib/view/root.py
@@ -1,0 +1,19 @@
+from tkinter import *
+
+from lib.view.pages.main import MainPage
+
+PROGRAM_NAME: str = "discoxy"
+
+
+class Root:
+    def __init__(self):
+        self.window = Tk()
+        self.window.title(PROGRAM_NAME)
+        self.window.geometry("420x210")
+        self.window.grid_rowconfigure(0, weight=1)
+        self.window.grid_columnconfigure(0, weight=1)
+
+        main_page = MainPage(self.window)
+        main_page.frame.tkraise()
+
+        self.window.mainloop()
